% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kstemind_tools.R
\name{calculate_ssm}
\alias{calculate_ssm}
\title{Calculate Stemmer Similarity Metrics (SSM)}
\usage{
calculate_ssm(output_stem_a, output_stem_b, original_words)
}
\arguments{
\item{output_stem_a}{an array of output stemmer A}

\item{output_stem_b}{an array of output stemmer B}

\item{original_words}{an array of original word tested in those two stemmer}
}
\value{
list value of SSM in range between 0-100 and df difference from both stemmer
}
\description{
Calculate Stemmer Similarity Metrics (SSM)
}
