% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kstemind_tools.R
\name{chunk_text}
\alias{chunk_text}
\title{Chunking Tokens of Word}
\usage{
chunk_text(text, by_size = F, chunk_size = 100)
}
\arguments{
\item{text}{long string}

\item{chunk_size}{(default = 100)}

\item{text_id}{(default = "teks")}
}
\value{
list of chunked tokenized text
}
\description{
Chunking Tokens of Word
}
