% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kstemind_tools.R
\name{chunk_tokenized_text}
\alias{chunk_tokenized_text}
\title{Chunking Tokens of Word}
\usage{
chunk_tokenized_text(tokenized_text, chunk_size = 100, text_id = "teks")
}
\arguments{
\item{tokenized_text}{array of string}

\item{chunk_size}{(default = 100)}

\item{text_id}{(default = "teks")}
}
\value{
list of chunked tokenized text
}
\description{
Chunking Tokens of Word
}
