% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kstemind_load_data.R
\name{kstemind_load_data}
\alias{kstemind_load_data}
\title{kstemind_load_data}
\usage{
kstemind_load_data(data)
}
\arguments{
\item{data}{string: the name of the data. Other than the specified option below will return NULL.
options : ("kamus_sastrawi", "kamus_nazief", "kamus_hidayat").
             ("tabel_visitors_cs", "tabel_visitors_ecs", "tabel_level_incremental"),
         ("aturan_nondeterministik")}
}
\value{
primary dataset in hashmap or data.frame
}
\description{
kstemind_load_data
}
\examples{
my_data <- kstemind_load_data("kamus_sastrawi")

}
