% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemmer_cs.R
\docType{class}
\name{stemmer_cs}
\alias{stemmer_cs}
\title{Stemmer CS - implementation of Confix Stripping Stemmer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
stemmer_cs
}
\description{
Stemmer CS - implementation of Confix Stripping Stemmer
}
\keyword{datasets}
