% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemmer_ecs.R
\docType{class}
\name{stemmer_ecs}
\alias{stemmer_ecs}
\title{Stemmer ECS - implementation of Enhanced CS Stemmer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
stemmer_ecs
}
\description{
Stemmer ECS - implementation of Enhanced CS Stemmer
}
\keyword{datasets}
