% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemmer_incremental.R
\docType{class}
\name{stemmer_incremental}
\alias{stemmer_incremental}
\title{Stemmer Incremental - implementation of Hidayat's Incremental Stemmer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
stemmer_incremental
}
\description{
Stemmer Incremental - implementation of Hidayat's Incremental Stemmer
}
\keyword{datasets}
