% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemmer_mecs.R
\docType{class}
\name{stemmer_mecs}
\alias{stemmer_mecs}
\title{Stemmer MECS - implementation of Modified ECS Stemmer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
stemmer_mecs
}
\description{
Stemmer MECS - implementation of Modified ECS Stemmer
}
\keyword{datasets}
