% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemmer_nazief_andriani.R
\docType{class}
\name{stemmer_nazief_andriani}
\alias{stemmer_nazief_andriani}
\title{Stemmer Nazief Andriani implementation of Nazief and Andriani's Porter Stemmer}
\format{An \code{\link{R6Class}} object.}
\usage{
stemmer_nazief_andriani
}
\value{
Object of \code{\link{R6Class}} with methods for stemming with Nazief and Andriani algorithm
}
\description{
Stemmer Nazief Andriani implementation of Nazief and Andriani's Porter Stemmer
}
\section{Methods}{


\describe{
   \item{\code{stem_word(word)}}{
     stemming a single word into its stem.
   }
}
}

\examples{

}
\keyword{stemmer}
