% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemmer_nondeterministic.R
\docType{class}
\name{stemmer_nondeterministic}
\alias{stemmer_nondeterministic}
\title{Stemmer Non-Deterministic - implementation of Non-Deterministic Stemmer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
stemmer_nondeterministic
}
\description{
Stemmer Non-Deterministic - implementation of Non-Deterministic Stemmer
}
\keyword{datasets}
