% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemmer_sastrawi.R
\docType{class}
\name{stemmer_sastrawi}
\alias{stemmer_sastrawi}
\title{Stemmer Sastrawi - Porting of the most popular bahasa Indonesia Stemmer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
stemmer_sastrawi
}
\description{
Stemmer Sastrawi - Porting of the most popular bahasa Indonesia Stemmer
}
\keyword{datasets}
