% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemmer_tala_porter.R
\docType{class}
\name{stemmer_tala_porter}
\alias{stemmer_tala_porter}
\title{Stemmer Tala Porter - implementation of Tala's Porter Stemmer}
\format{An \code{\link{R6Class}} object.}
\usage{
stemmer_tala_porter
}
\value{
Object of \code{\link{R6Class}} with methods for stemming with Tala's Porter algorithm
}
\description{
Stemmer Tala Porter - implementation of Tala's Porter Stemmer
}
\section{Methods}{


\describe{
   \item{\code{stem_word(word)}}{
     stemming a single word into its stem.
   }

   \item{\code{affix_particle}}{}
   \item{\code{affix_possesive}}{}
   \item{\code{affix_first_order_prefix}}{}
   \item{\code{affix_second_order_prefix}}{}
   \item{\code{affix_suffix}}{}
   \item{\code{special_words}}{}
}
}

\examples{

}
\keyword{data}
