% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kstemind_tools.R
\name{stopwords_removal}
\alias{stopwords_removal}
\title{Stopwords Removal}
\usage{
stopwords_removal(tokenized_text)
}
\arguments{
\item{tokenized_text}{an array of tokens}
}
\value{
filtered array of tokens
}
\description{
Stopwords Removal
}
