% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kstemind_tools.R
\name{tokenize_indonesian}
\alias{tokenize_indonesian}
\title{Tokenizer for Indonesian text}
\usage{
tokenize_indonesian(text, do_clean = FALSE)
}
\arguments{
\item{text}{string}

\item{do_clean}{boolean}
}
\value{
array of tokenized string
}
\description{
Tokenizer for Indonesian text
}
