% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemmer_cs_extensions.R
\docType{class}
\name{visitors_cs}
\alias{visitors_cs}
\title{Visitors CS - Internal Function for Prefixes Removal in Confix Stripping Stemmer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
visitors_cs
}
\description{
Visitors CS - Internal Function for Prefixes Removal in Confix Stripping Stemmer
}
\keyword{datasets}
