% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemmer_ecs_extensions.R
\docType{class}
\name{visitors_ecs}
\alias{visitors_ecs}
\title{Visitors ECS - Internal Function for Prefixes Removal in Enhanced CS Stemmer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
visitors_ecs
}
\description{
Visitors ECS - Internal Function for Prefixes Removal in Enhanced CS Stemmer
}
\keyword{datasets}
